moved inspect_string_omit_overlap into separate file
